requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
cfgv==3.1.0
click==7.1.2
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-favicon==0.1.3
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
drf-extensions==0.6.0
filelock==3.0.12
flake8==3.8.1
gunicorn==20.0.4
identify==1.4.16
mccabe==0.6.1
nodeenv==1.3.5
pathspec==0.8.0
pre-commit==2.4.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.21
whitenoise==5.1.0
